improved Netflix fullscreen mode and mouse hiding
* UNDNetflixPlayer.m

Previously a simulated mouse click was used to enter full-screen mode, but this
was unreliable depending on the video aspect ratio and the screen size. By
simulating a press of the 'f' key we can more reliably activate full-screen.

The previous method of hiding the mouse by moving it to a screen corner was
problematic because 1) it might effect "Active Screen Corners" and 2) it would
interfere with attempted user input. Now we hide/show the cursor depending on
use idle time.

* UNDNetflixKeyboardDevice.m: (new file)

Added a new keyboard mapping specifically for the Netflix player.