abstract media asset storage
Each asset type is generated from a dictionary description that indicates what
type of media it is. Each media service has a new corresponding asset provider.
The various dialogs for adding new assets are updated to use the new providers.
A single asset factory handles a registry of the asset providers, which must be
registered. The preference system is modified to keep asset descriptions
instead of URLs and titles.